be better about cleaning up screen sessions

+++ check.sh
@@ -28,14 +28,15 @@ else
 fi
 
 debug "building..."
-./allbuild.sh > /dev/null
+./build.sh > /dev/null
 debug "done building"
 
 WRONGS=""
 RIGHTS=""
 
 for Q in $QUESTIONS ; do
-  EXEC="$Q/build/$Q"
+  BUILD_DIR="$Q/cmake-build"
+  EXEC="$BUILD_DIR/$Q"
   info "evaluating $EXEC"
   INFILE="$Q/input.txt"
   EXPECTEDFILE="$Q/expected-output.txt"
@@ -43,10 +44,17 @@ for Q in $QUESTIONS ; do
     echo "file missing: $EXPECTEDFILE" >&2
     exit 1
   fi
-  OUTFILE="$Q/build/actual.txt"
+  OUTFILE="$BUILD_DIR/actual.txt"
+  touch "$OUTFILE"
+  truncate --size=0 "$OUTFILE"
   if [ -f "$INFILE" ] ; then
+    if [ "$(wc -l < "$INFILE")" == "0" ] ; then
+      echo "must contain newline: $INFILE" <&2
+      exit 1
+    fi
     rm -rf ./screenlog.*
-    SESSIONNAME=$(basename "$Q")
+    TEMPFILE=$(mktemp --tmpdir="$BUILD_DIR" sess_XXXXXXXXX)
+    SESSIONNAME=$(basename "$TEMPFILE")
     screen -L -S "$SESSIONNAME" -d -m "$EXEC"
     debug "sleeping ${PAUSE} after opening screen session $SESSIONNAME"
     sleep "$PAUSE"
@@ -57,14 +65,14 @@ for Q in $QUESTIONS ; do
       debug "sleeping ${PAUSE} after entering $line"
       sleep "$PAUSE"
     done < "$INFILE"
+    screen -S "$SESSIONNAME" -X quit >/dev/null || true 
+    rm "$TEMPFILE" -f
     dos2unix -q -n screenlog.0 "$OUTFILE"
     rm screenlog.0
   else
     echo "$Q: no input file" >&2
     "$EXEC" > "$OUTFILE"
   fi
-  #ACTUAL=$(cat "$OUTFILE")
-  #EXPECTED=$(cat "$EXPECTEDFILE")
   
   if diff "$EXPECTEDFILE" "$OUTFILE" ; then
     RIGHTS="$Q $RIGHTS"
@@ -75,7 +83,7 @@ done
 
 if [ -n "$WRONGS" ] ; then
   echo "wrong: $WRONGS" >&2
-  exit 1
+  exit 2
 else
   echo "gravy: $RIGHTS"
 fi
